Move SolarSystemDemo to ExampleBrowser
