+++ Source/HelixToolkit.Wpf.sln
@@ -17,8 +17,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HelixToolkit.Wpf", "HelixTo
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PanoramaDemo", "Examples\WPF\PanoramaDemo\PanoramaDemo.csproj", "{E68C0BD5-316A-4546-9FFC-9722350E4467}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SolarsystemDemo", "Examples\WPF\SolarsystemDemo\SolarsystemDemo.csproj", "{DFE51472-60CD-4F9C-8E52-47B250F58850}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tools", "Tools", "{4EC16C60-5AFA-4854-82D9-FB1BF92C3812}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BatchRender", "Tools\BatchRender\BatchRender.csproj", "{0DA2D747-B3FB-40A6-B8FF-0F8949A3E1BC}"
@@ -108,16 +106,6 @@ Global
 		{E68C0BD5-316A-4546-9FFC-9722350E4467}.Release|Mixed Platforms.Build.0 = Release|x86
 		{E68C0BD5-316A-4546-9FFC-9722350E4467}.Release|x86.ActiveCfg = Release|x86
 		{E68C0BD5-316A-4546-9FFC-9722350E4467}.Release|x86.Build.0 = Release|x86
-		{DFE51472-60CD-4F9C-8E52-47B250F58850}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{DFE51472-60CD-4F9C-8E52-47B250F58850}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{DFE51472-60CD-4F9C-8E52-47B250F58850}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{DFE51472-60CD-4F9C-8E52-47B250F58850}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{DFE51472-60CD-4F9C-8E52-47B250F58850}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{DFE51472-60CD-4F9C-8E52-47B250F58850}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{DFE51472-60CD-4F9C-8E52-47B250F58850}.Release|Any CPU.Build.0 = Release|Any CPU
-		{DFE51472-60CD-4F9C-8E52-47B250F58850}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{DFE51472-60CD-4F9C-8E52-47B250F58850}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{DFE51472-60CD-4F9C-8E52-47B250F58850}.Release|x86.ActiveCfg = Release|Any CPU
 		{0DA2D747-B3FB-40A6-B8FF-0F8949A3E1BC}.Debug|Any CPU.ActiveCfg = Debug|x86
 		{0DA2D747-B3FB-40A6-B8FF-0F8949A3E1BC}.Debug|Mixed Platforms.ActiveCfg = Debug|x86
 		{0DA2D747-B3FB-40A6-B8FF-0F8949A3E1BC}.Debug|Mixed Platforms.Build.0 = Debug|x86
@@ -197,7 +185,6 @@ Global
 		{B262B3C4-B4F7-4CD0-9055-62B271856078} = {848BE1AD-560A-4419-BD41-994A1A1CCF1F}
 		{CD249FE2-999C-4AC3-810D-C53ACB2A7DF3} = {848BE1AD-560A-4419-BD41-994A1A1CCF1F}
 		{E68C0BD5-316A-4546-9FFC-9722350E4467} = {848BE1AD-560A-4419-BD41-994A1A1CCF1F}
-		{DFE51472-60CD-4F9C-8E52-47B250F58850} = {848BE1AD-560A-4419-BD41-994A1A1CCF1F}
 		{0DA2D747-B3FB-40A6-B8FF-0F8949A3E1BC} = {4EC16C60-5AFA-4854-82D9-FB1BF92C3812}
 		{47FC8A4C-86B6-4D9D-BE10-EE74F8B992BB} = {848BE1AD-560A-4419-BD41-994A1A1CCF1F}
 		{991A4C2B-D830-41CB-A810-047ADCF69219} = {848BE1AD-560A-4419-BD41-994A1A1CCF1F}
